requirements.txt
pyside6
python-can
opencv-python
